3° commit

+++ Sorteio/src/main/java/mvc/SorteioNumeroControle.java
@@ -17,10 +17,10 @@ public class SorteioNumeroControle extends HttpServlet{
       throws ServletException, IOException {
 
     String paramNumero1 = req.getParameter("numero1");
-    int numero1 = paramNumero1 == null ? 0.0 : Integer.parseInt(paramNumero1);
+    int numero1 = (int) (paramNumero1 == null ? 0.0 : Integer.parseInt(paramNumero1));
     
     String paramNumero2 = req.getParameter("numero2");
-    int numero2 = paramNumero2 == null ? 0.0 : Integer.parseInt(paramNumero2);
+    int numero2 = (int) (paramNumero2 == null ? 0.0 : Integer.parseInt(paramNumero2));
 
     SorteioNumeroModel cSorteio = new SorteioNumeroModel();
     cSorteio.getNumero1(numero1);
